# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VoxChatRequestMessagesItem = {
  role: string;
  content: string;
};

export interface VoxChatRequest {
  messages: VoxChatRequestMessagesItem[];
}

export interface VoxTranscribeRequest {
  /** Base64-encoded audio data */
  audio: string;
  /** Audio format (m4a, wav, mp3, etc.) */
  format?: string;
}

export interface VoxTranscribeResponse {
  text: string;
}

export interface VoxSpeakRequest {
  text: string;
  /** Voice ID: alloy, echo, fable, onyx, nova, shimmer */
  voice?: string;
}

export interface VoxSpeakResponse {
  /** Base64-encoded audio */
  audio: string;
  format: string;
}

export interface VoxError {
  error: string;
}