# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Chat with VOX AI (streaming SSE)
 */
export const VoxChatBody = zod.object({
  "messages": zod.array(zod.object({
  "role": zod.string(),
  "content": zod.string()
}))
})


/**
 * @summary Transcribe audio to text
 */
export const VoxTranscribeBody = zod.object({
  "audio": zod.string().describe('Base64-encoded audio data'),
  "format": zod.string().optional().describe('Audio format (m4a, wav, mp3, etc.)')
})

export const VoxTranscribeResponse = zod.object({
  "text": zod.string()
})


/**
 * @summary Convert text to speech
 */
export const VoxSpeakBody = zod.object({
  "text": zod.string(),
  "voice": zod.string().optional().describe('Voice ID: alloy, echo, fable, onyx, nova, shimmer')
})

export const VoxSpeakResponse = zod.object({
  "audio": zod.string().describe('Base64-encoded audio'),
  "format": zod.string()
})