# lib/api-zod/src/generated/types/voxChatRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VoxChatRequestMessagesItem } from './voxChatRequestMessagesItem';

export interface VoxChatRequest {
  messages: VoxChatRequestMessagesItem[];
}