# lib/api-zod/src/generated/types/voxSpeakRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VoxSpeakRequest {
  text: string;
  /** Voice ID: alloy, echo, fable, onyx, nova, shimmer */
  voice?: string;
}