# lib/api-zod/src/generated/types/voxTranscribeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VoxTranscribeRequest {
  /** Base64-encoded audio data */
  audio: string;
  /** Audio format (m4a, wav, mp3, etc.) */
  format?: string;
}